Removing obsolete inventory fields and disabling tests
