--- plugins/woocommerce-admin/client/products/sections/product-inventory-section/index.ts
@@ -1,2 +1,0 @@
-export * from './product-inventory-section';
-export * from './utils';
